requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# Telegram
python-telegram-bot>=21.0

# Database
aiosqlite>=0.19.0

# Voice transcription
openai-whisper>=20231117

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Web scraping (URL title extraction)
httpx>=0.27.0
beautifulsoup4>=4.12.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
